+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Progress bar on training, validation and evaluation.
 - Model inference on image file.
 - Model summary in train, evaluate and export program.
+- Model fine-tuning.
  
 
 ### Changed